refactor(examples): unify, simplify, group examples
Squashed from 6 commits:
- extract shared helpers into examples/common
- make WebAuthn examples JSON-first
- rename for naming consistency
- group into ceremony/features/management subdirs
- update README example table for JSON-first WebAuthn and grouped layout
- drop unit let-binding in change_pin_hid
