+++ libwebauthn/examples/ceremony/webauthn_hid.rs
@@ -1,76 +1,23 @@
 use std::error::Error;
-use std::io::{self, Write};
 use std::time::Duration;
-
-use libwebauthn::UvUpdate;
-use text_io::read;
-use tokio::sync::broadcast::Receiver;
-use tracing_subscriber::{self, EnvFilter};
 
 use libwebauthn::ops::webauthn::{
     DatFilePublicSuffixList, GetAssertionRequest, JsonFormat, MakeCredentialRequest, RequestOrigin,
     WebAuthnIDL as _, WebAuthnIDLResponse as _,
 };
-use libwebauthn::pin::PinRequestReason;
 use libwebauthn::proto::ctap2::Ctap2PublicKeyCredentialDescriptor;
 use libwebauthn::transport::hid::list_devices;
 use libwebauthn::transport::{Channel as _, Device};
-use libwebauthn::webauthn::{Error as WebAuthnError, WebAuthn};
+use libwebauthn::webauthn::WebAuthn;
+
+#[path = "../common/mod.rs"]
+mod common;
 
 const TIMEOUT: Duration = Duration::from_secs(10);
 
-fn setup_logging() {
-    tracing_subscriber::fmt()
-        .with_env_filter(EnvFilter::from_default_env())
-        .without_time()
-        .init();
-}
-
-async fn handle_updates(mut state_recv: Receiver<UvUpdate>) {
-    while let Ok(update) = state_recv.recv().await {
-        match update {
-            UvUpdate::PresenceRequired => println!("Please touch your device!"),
-            // Shouldn't really happen, as we don't use UV required
-            UvUpdate::PinNotSet(_) => println!("Pin not set for your device!"),
-            UvUpdate::UvRetry { attempts_left } => {
-                print!("UV failed.");
-                if let Some(attempts_left) = attempts_left {
-                    print!(" You have {attempts_left} attempts left.");
-                }
-            }
-            UvUpdate::PinRequired(update) => {
-                let mut attempts_str = String::new();
-                if let Some(attempts) = update.attempts_left {
-                    attempts_str = format!(". You have {attempts} attempts left!");
-                };
-
-                match update.reason {
-                    PinRequestReason::RelyingPartyRequest => println!("RP required a PIN."),
-                    PinRequestReason::AuthenticatorPolicy => {
-                        println!("Your device requires a PIN.")
-                    }
-                    PinRequestReason::FallbackFromUV => {
-                        println!("UV failed too often and is blocked. Falling back to PIN.")
-                    }
-                }
-                print!("PIN: Please enter the PIN for your authenticator{attempts_str}: ");
-                io::stdout().flush().unwrap();
-                let pin_raw: String = read!("{}\n");
-
-                if pin_raw.is_empty() {
-                    println!("PIN: No PIN provided, cancelling operation.");
-                    update.cancel();
-                } else {
-                    let _ = update.send_pin(&pin_raw);
-                }
-            }
-        }
-    }
-}
-
 #[tokio::main]
 pub async fn main() -> Result<(), Box<dyn Error>> {
-    setup_logging();
+    common::setup_logging();
 
     let devices = list_devices().await.unwrap();
     println!("Devices found: {:?}", devices);
@@ -118,29 +65,13 @@ pub async fn main() -> Result<(), Box<dyn Error>> {
         );
 
         let state_recv = channel.get_ux_update_receiver();
-        tokio::spawn(handle_updates(state_recv));
+        tokio::spawn(common::handle_uv_updates(state_recv));
 
-        let response = loop {
-            match channel
-                .webauthn_make_credential(&make_credentials_request)
-                .await
-            {
-                Ok(response) => break Ok(response),
-                Err(WebAuthnError::Ctap(ctap_error)) => {
-                    if ctap_error.is_retryable_user_error() {
-                        println!("Oops, try again! Error: {}", ctap_error);
-                        continue;
-                    }
-                    break Err(WebAuthnError::Ctap(ctap_error));
-                }
-                Err(err) => break Err(err),
-            };
-        }
-        .unwrap();
+        let response =
+            retry_user_errors!(channel.webauthn_make_credential(&make_credentials_request))
+                .unwrap();
         println!("WebAuthn MakeCredential response: {:?}", response);
 
-        // Serialize the response back to JSON using the original request as context.
-        // The request contains the client_data_json bytes that were built during parsing.
         match response.to_json_string(&make_credentials_request, JsonFormat::Prettified) {
             Ok(response_json) => {
                 println!(
@@ -174,23 +105,9 @@ pub async fn main() -> Result<(), Box<dyn Error>> {
                 .expect("Failed to parse request JSON");
         println!("WebAuthn GetAssertion request: {:?}", get_assertion);
 
-        let response = loop {
-            match channel.webauthn_get_assertion(&get_assertion).await {
-                Ok(response) => break Ok(response),
-                Err(WebAuthnError::Ctap(ctap_error)) => {
-                    if ctap_error.is_retryable_user_error() {
-                        println!("Oops, try again! Error: {}", ctap_error);
-                        continue;
-                    }
-                    break Err(WebAuthnError::Ctap(ctap_error));
-                }
-                Err(err) => break Err(err),
-            };
-        }
-        .unwrap();
+        let response = retry_user_errors!(channel.webauthn_get_assertion(&get_assertion)).unwrap();
         println!("WebAuthn GetAssertion response: {:?}", response);
 
-        // Serialize the response back to JSON using the original request as context.
         for assertion in &response.assertions {
             match assertion.to_json_string(&get_assertion, JsonFormat::Prettified) {
                 Ok(assertion_json) => {
