+++ libwebauthn/examples/features/webauthn_preflight_hid.rs
@@ -1,19 +1,13 @@
 use std::error::Error;
-use std::io::{self, Write};
 use std::time::Duration;
 
-use libwebauthn::UvUpdate;
 use rand::{thread_rng, Rng};
 use serde_bytes::ByteBuf;
-use text_io::read;
-use tokio::sync::broadcast::Receiver;
-use tracing_subscriber::{self, EnvFilter};
 
 use libwebauthn::ops::webauthn::{
     GetAssertionRequest, GetAssertionRequestExtensions, GetAssertionResponse,
     MakeCredentialRequest, ResidentKeyRequirement, UserVerificationRequirement,
 };
-use libwebauthn::pin::PinRequestReason;
 use libwebauthn::proto::ctap2::{
     Ctap2CredentialType, Ctap2PublicKeyCredentialDescriptor, Ctap2PublicKeyCredentialRpEntity,
     Ctap2PublicKeyCredentialType, Ctap2PublicKeyCredentialUserEntity,
@@ -22,60 +16,14 @@ use libwebauthn::transport::hid::list_devices;
 use libwebauthn::transport::{Channel, Device};
 use libwebauthn::webauthn::{CtapError, Error as WebAuthnError, WebAuthn};
 
+#[path = "../common/mod.rs"]
+mod common;
+
 const TIMEOUT: Duration = Duration::from_secs(10);
-
-fn setup_logging() {
-    tracing_subscriber::fmt()
-        .with_env_filter(EnvFilter::from_default_env())
-        .without_time()
-        .init();
-}
-
-async fn handle_updates(mut state_recv: Receiver<UvUpdate>) {
-    while let Ok(update) = state_recv.recv().await {
-        match update {
-            UvUpdate::PresenceRequired => println!("Please touch your device!"),
-            // Shouldn't really happen, as we don't use UV required
-            UvUpdate::PinNotSet(_) => println!("Pin not set for your device!"),
-            UvUpdate::UvRetry { attempts_left } => {
-                print!("UV failed.");
-                if let Some(attempts_left) = attempts_left {
-                    print!(" You have {attempts_left} attempts left.");
-                }
-            }
-            UvUpdate::PinRequired(update) => {
-                let mut attempts_str = String::new();
-                if let Some(attempts) = update.attempts_left {
-                    attempts_str = format!(". You have {attempts} attempts left!");
-                };
-
-                match update.reason {
-                    PinRequestReason::RelyingPartyRequest => println!("RP required a PIN."),
-                    PinRequestReason::AuthenticatorPolicy => {
-                        println!("Your device requires a PIN.")
-                    }
-                    PinRequestReason::FallbackFromUV => {
-                        println!("UV failed too often and is blocked. Falling back to PIN.")
-                    }
-                }
-                print!("PIN: Please enter the PIN for your authenticator{attempts_str}: ");
-                io::stdout().flush().unwrap();
-                let pin_raw: String = read!("{}\n");
-
-                if pin_raw.is_empty() {
-                    println!("PIN: No PIN provided, cancelling operation.");
-                    update.cancel();
-                } else {
-                    let _ = update.send_pin(&pin_raw);
-                }
-            }
-        }
-    }
-}
 
 #[tokio::main]
 pub async fn main() -> Result<(), Box<dyn Error>> {
-    setup_logging();
+    common::setup_logging();
 
     println!("-------------------------------------------------------");
     println!("Run this test with RUST_LOG=libwebauthn::proto::ctap2::preflight=info to verify the outputs");
@@ -91,7 +39,7 @@ pub async fn main() -> Result<(), Box<dyn Error>> {
         channel.wink(TIMEOUT).await?;
 
         let state_recv = channel.get_ux_update_receiver();
-        tokio::spawn(handle_updates(state_recv));
+        tokio::spawn(common::handle_uv_updates(state_recv));
 
         println!("Make credential with exclude_list: None. Should do nothing in preflight and return a credential:");
         let res = make_credential_call(&mut channel, &user_id, None).await;
@@ -175,23 +123,8 @@ async fn make_credential_call(
         timeout: TIMEOUT,
     };
 
-    let response = loop {
-        match channel
-            .webauthn_make_credential(&make_credentials_request)
-            .await
-        {
-            Ok(response) => break Ok(response),
-            Err(WebAuthnError::Ctap(ctap_error)) => {
-                if ctap_error.is_retryable_user_error() {
-                    println!("Oops, try again! Error: {}", ctap_error);
-                    continue;
-                }
-                break Err(WebAuthnError::Ctap(ctap_error));
-            }
-            Err(err) => break Err(err),
-        };
-    };
-    response.map(|x| (&x.authenticator_data).try_into().unwrap())
+    retry_user_errors!(channel.webauthn_make_credential(&make_credentials_request))
+        .map(|x| (&x.authenticator_data).try_into().unwrap())
 }
 
 async fn get_assertion_call(
@@ -210,20 +143,7 @@ async fn get_assertion_call(
         timeout: TIMEOUT,
     };
 
-    let response = loop {
-        match channel.webauthn_get_assertion(&get_assertion).await {
-            Ok(response) => break Ok(response),
-            Err(WebAuthnError::Ctap(ctap_error)) => {
-                if ctap_error.is_retryable_user_error() {
-                    println!("Oops, try again! Error: {}", ctap_error);
-                    continue;
-                }
-                break Err(WebAuthnError::Ctap(ctap_error));
-            }
-            Err(err) => break Err(err),
-        };
-    };
-    response
+    retry_user_errors!(channel.webauthn_get_assertion(&get_assertion))
 }
 
 fn create_credential(id: &[u8]) -> Ctap2PublicKeyCredentialDescriptor {
