+++ libwebauthn/examples/features/webauthn_prf_hid.rs
@@ -1,22 +1,16 @@
 use std::collections::HashMap;
 use std::convert::TryInto;
 use std::error::Error;
-use std::io::{self, Write};
 use std::time::Duration;
 
 use libwebauthn::transport::hid::channel::HidChannel;
-use libwebauthn::UvUpdate;
 use rand::{thread_rng, Rng};
-use text_io::read;
-use tokio::sync::broadcast::Receiver;
-use tracing_subscriber::{self, EnvFilter};
 
 use libwebauthn::ops::webauthn::{
     GetAssertionRequest, GetAssertionRequestExtensions, MakeCredentialPrfInput,
     MakeCredentialRequest, MakeCredentialsRequestExtensions, PRFValue, PrfInput,
     ResidentKeyRequirement, UserVerificationRequirement,
 };
-use libwebauthn::pin::PinRequestReason;
 use libwebauthn::proto::ctap2::{
     Ctap2CredentialType, Ctap2PublicKeyCredentialDescriptor, Ctap2PublicKeyCredentialRpEntity,
     Ctap2PublicKeyCredentialUserEntity,
@@ -25,60 +19,14 @@ use libwebauthn::transport::hid::list_devices;
 use libwebauthn::transport::{Channel as _, Device};
 use libwebauthn::webauthn::{Error as WebAuthnError, PlatformError, WebAuthn};
 
+#[path = "../common/mod.rs"]
+mod common;
+
 const TIMEOUT: Duration = Duration::from_secs(10);
-
-fn setup_logging() {
-    tracing_subscriber::fmt()
-        .with_env_filter(EnvFilter::from_default_env())
-        .without_time()
-        .init();
-}
-
-async fn handle_updates(mut state_recv: Receiver<UvUpdate>) {
-    while let Ok(update) = state_recv.recv().await {
-        match update {
-            UvUpdate::PresenceRequired => println!("Please touch your device!"),
-            // Shouldn't really happen, as we don't use UV required
-            UvUpdate::PinNotSet(_) => println!("Pin not set for your device!"),
-            UvUpdate::UvRetry { attempts_left } => {
-                print!("UV failed.");
-                if let Some(attempts_left) = attempts_left {
-                    print!(" You have {attempts_left} attempts left.");
-                }
-            }
-            UvUpdate::PinRequired(update) => {
-                let mut attempts_str = String::new();
-                if let Some(attempts) = update.attempts_left {
-                    attempts_str = format!(". You have {attempts} attempts left!");
-                };
-
-                match update.reason {
-                    PinRequestReason::RelyingPartyRequest => println!("RP required a PIN."),
-                    PinRequestReason::AuthenticatorPolicy => {
-                        println!("Your device requires a PIN.")
-                    }
-                    PinRequestReason::FallbackFromUV => {
-                        println!("UV failed too often and is blocked. Falling back to PIN.")
-                    }
-                }
-                print!("PIN: Please enter the PIN for your authenticator{attempts_str}: ");
-                io::stdout().flush().unwrap();
-                let pin_raw: String = read!("{}\n");
-
-                if pin_raw.is_empty() {
-                    println!("PIN: No PIN provided, cancelling operation.");
-                    update.cancel();
-                } else {
-                    let _ = update.send_pin(&pin_raw);
-                }
-            }
-        }
-    }
-}
 
 #[tokio::main]
 pub async fn main() -> Result<(), Box<dyn Error>> {
-    setup_logging();
+    common::setup_logging();
 
     let devices = list_devices().await.unwrap();
     println!("Devices found: {:?}", devices);
@@ -97,9 +45,8 @@ pub async fn main() -> Result<(), Box<dyn Error>> {
         channel.wink(TIMEOUT).await?;
 
         let state_recv = channel.get_ux_update_receiver();
-        tokio::spawn(handle_updates(state_recv));
+        tokio::spawn(common::handle_uv_updates(state_recv));
 
-        // Make Credentials ceremony
         let make_credentials_request = MakeCredentialRequest {
             challenge: Vec::from(challenge),
             origin: "example.org".to_owned(),
@@ -114,23 +61,9 @@ pub async fn main() -> Result<(), Box<dyn Error>> {
             timeout: TIMEOUT,
         };
 
-        let response = loop {
-            match channel
-                .webauthn_make_credential(&make_credentials_request)
-                .await
-            {
-                Ok(response) => break Ok(response),
-                Err(WebAuthnError::Ctap(ctap_error)) => {
-                    if ctap_error.is_retryable_user_error() {
-                        println!("Oops, try again! Error: {}", ctap_error);
-                        continue;
-                    }
-                    break Err(WebAuthnError::Ctap(ctap_error));
-                }
-                Err(err) => break Err(err),
-            };
-        }
-        .unwrap();
+        let response =
+            retry_user_errors!(channel.webauthn_make_credential(&make_credentials_request))
+                .unwrap();
 
         println!(
             "WebAuthn MakeCredential extensions: {:?}",
@@ -141,8 +74,6 @@ pub async fn main() -> Result<(), Box<dyn Error>> {
             (&response.authenticator_data).try_into().unwrap();
 
         // Test 1: eval_by_credential with the cred_id we got
-        let eval = None;
-
         let mut eval_by_credential = HashMap::new();
         eval_by_credential.insert(
             base64_url::encode(&credential.id),
@@ -151,25 +82,19 @@ pub async fn main() -> Result<(), Box<dyn Error>> {
                 second: None,
             },
         );
-        let prf = PrfInput {
-            eval,
-            eval_by_credential,
-        };
         run_success_test(
             &mut channel,
             &credential,
             &challenge,
-            prf,
+            PrfInput {
+                eval: None,
+                eval_by_credential,
+            },
             "eval_by_credential only",
         )
         .await;
 
         // Test 2: eval and eval_with_credential with cred_id we got
-        let eval = Some(PRFValue {
-            first: [2; 32],
-            second: None,
-        });
-
         let mut eval_by_credential = HashMap::new();
         eval_by_credential.insert(
             base64_url::encode(&credential.id),
@@ -178,38 +103,38 @@ pub async fn main() -> Result<(), Box<dyn Error>> {
                 second: None,
             },
         );
-        let prf = PrfInput {
-            eval,
-            eval_by_credential,
-        };
         run_success_test(
             &mut channel,
             &credential,
             &challenge,
-            prf,
+            PrfInput {
+                eval: Some(PRFValue {
+                    first: [2; 32],
+                    second: None,
+                }),
+                eval_by_credential,
+            },
             "eval and eval_by_credential",
         )
         .await;
 
         // Test 3: eval only
-        let eval = Some(PRFValue {
-            first: [1; 32],
-            second: None,
-        });
-
-        let eval_by_credential = HashMap::new();
-        let prf = PrfInput {
-            eval,
-            eval_by_credential,
-        };
-        run_success_test(&mut channel, &credential, &challenge, prf, "eval only").await;
+        run_success_test(
+            &mut channel,
+            &credential,
+            &challenge,
+            PrfInput {
+                eval: Some(PRFValue {
+                    first: [1; 32],
+                    second: None,
+                }),
+                eval_by_credential: HashMap::new(),
+            },
+            "eval only",
+        )
+        .await;
 
         // Test 4: eval and a full list of eval_by_credential
-        let eval = Some(PRFValue {
-            first: [2; 32],
-            second: None,
-        });
-
         let mut eval_by_credential = HashMap::new();
         eval_by_credential.insert(
             base64_url::encode(&[5; 54]),
@@ -239,25 +164,22 @@ pub async fn main() -> Result<(), Box<dyn Error>> {
                 second: None,
             },
         );
-        let prf = PrfInput {
-            eval,
-            eval_by_credential,
-        };
         run_success_test(
             &mut channel,
             &credential,
             &challenge,
-            prf,
+            PrfInput {
+                eval: Some(PRFValue {
+                    first: [2; 32],
+                    second: None,
+                }),
+                eval_by_credential,
+            },
             "eval and full list of eval_by_credential",
         )
         .await;
 
         // Test 5: eval and non-fitting list of eval_by_credential
-        let eval = Some(PRFValue {
-            first: [1; 32],
-            second: None,
-        });
-
         let mut eval_by_credential = HashMap::new();
         eval_by_credential.insert(
             base64_url::encode(&[5; 54]),
@@ -280,22 +202,22 @@ pub async fn main() -> Result<(), Box<dyn Error>> {
                 second: Some([8; 32]),
             },
         );
-        let prf = PrfInput {
-            eval,
-            eval_by_credential,
-        };
         run_success_test(
             &mut channel,
             &credential,
             &challenge,
-            prf,
+            PrfInput {
+                eval: Some(PRFValue {
+                    first: [1; 32],
+                    second: None,
+                }),
+                eval_by_credential,
+            },
             "eval and non-fitting list of eval_by_credential",
         )
         .await;
 
         // Test 6: no eval and non-fitting list of eval_by_credential
-        let eval = None;
-
         let mut eval_by_credential = HashMap::new();
         eval_by_credential.insert(
             base64_url::encode(&[5; 54]),
@@ -318,25 +240,19 @@ pub async fn main() -> Result<(), Box<dyn Error>> {
                 second: Some([8; 32]),
             },
         );
-        let prf = PrfInput {
-            eval,
-            eval_by_credential,
-        };
         run_success_test(
             &mut channel,
             &credential,
             &challenge,
-            prf,
+            PrfInput {
+                eval: None,
+                eval_by_credential,
+            },
             "No eval and non-fitting list of eval_by_credential (should have no extension output)",
         )
         .await;
 
         // Test 7: Wrongly encoded credential_id
-        let eval = Some(PRFValue {
-            first: [2; 32],
-            second: None,
-        });
-
         let mut eval_by_credential = HashMap::new();
         eval_by_credential.insert(
             String::from("ÄöoLfwekldß^"),
@@ -345,22 +261,23 @@ pub async fn main() -> Result<(), Box<dyn Error>> {
                 second: None,
             },
         );
-        let prf = PrfInput {
-            eval,
-            eval_by_credential,
-        };
         run_failed_test(
             &mut channel,
             Some(&credential),
             &challenge,
-            prf,
+            PrfInput {
+                eval: Some(PRFValue {
+                    first: [2; 32],
+                    second: None,
+                }),
+                eval_by_credential,
+            },
             "Wrongly encoded credential_id",
             WebAuthnError::Platform(PlatformError::SyntaxError),
         )
         .await;
 
         // Test 8: Empty credential_id
-        let eval = None;
         let mut eval_by_credential = HashMap::new();
         eval_by_credential.insert(
             String::new(),
@@ -369,22 +286,20 @@ pub async fn main() -> Result<(), Box<dyn Error>> {
                 second: None,
             },
         );
-        let prf = PrfInput {
-            eval,
-            eval_by_credential,
-        };
         run_failed_test(
             &mut channel,
             Some(&credential),
             &challenge,
-            prf,
+            PrfInput {
+                eval: None,
+                eval_by_credential,
+            },
             "Empty credential_id",
             WebAuthnError::Platform(PlatformError::SyntaxError),
         )
         .await;
 
         // Test 9: Empty allow_list, set eval_by_credential
-        let eval = None;
         let mut eval_by_credential = HashMap::new();
         eval_by_credential.insert(
             String::new(),
@@ -393,15 +308,14 @@ pub async fn main() -> Result<(), Box<dyn Error>> {
                 second: None,
             },
         );
-        let prf = PrfInput {
-            eval,
-            eval_by_credential,
-        };
         run_failed_test(
             &mut channel,
             None,
             &challenge,
-            prf,
+            PrfInput {
+                eval: None,
+                eval_by_credential,
+            },
             "Empty allow_list, set eval_by_credential",
             WebAuthnError::Platform(PlatformError::NotSupported),
         )
@@ -431,20 +345,7 @@ async fn run_success_test(
         timeout: TIMEOUT,
     };
 
-    let response = loop {
-        match channel.webauthn_get_assertion(&get_assertion).await {
-            Ok(response) => break Ok(response),
-            Err(WebAuthnError::Ctap(ctap_error)) => {
-                if ctap_error.is_retryable_user_error() {
-                    println!("Oops, try again! Error: {}", ctap_error);
-                    continue;
-                }
-                break Err(WebAuthnError::Ctap(ctap_error));
-            }
-            Err(err) => break Err(err),
-        }
-    }
-    .unwrap();
+    let response = retry_user_errors!(channel.webauthn_get_assertion(&get_assertion)).unwrap();
     for (num, assertion) in response.assertions.iter().enumerate() {
         println!(
             "{num}. result of {printoutput}: {:?}",
@@ -475,19 +376,8 @@ async fn run_failed_test(
         timeout: TIMEOUT,
     };
 
-    let response: Result<(), libwebauthn::webauthn::Error> = loop {
-        match channel.webauthn_get_assertion(&get_assertion).await {
-            Ok(_) => panic!("Success, even though it should have errored out!"),
-            Err(WebAuthnError::Ctap(ctap_error)) => {
-                if ctap_error.is_retryable_user_error() {
-                    println!("Oops, try again! Error: {}", ctap_error);
-                    continue;
-                }
-                break Err(WebAuthnError::Ctap(ctap_error));
-            }
-            Err(err) => break Err(err),
-        }
-    };
+    let response = retry_user_errors!(channel.webauthn_get_assertion(&get_assertion))
+        .map(|_| panic!("Success, even though it should have errored out!"));
 
     assert_eq!(response, Err(expected_error), "{printoutput}:");
     println!("Success for test: {printoutput}")
