+++ libwebauthn/examples/management/authenticator_config_hid.rs
@@ -1,0 +1,148 @@
+use std::error::Error;
+use std::fmt::Display;
+use std::io::{self, Write};
+use std::time::Duration;
+
+use libwebauthn::management::AuthenticatorConfig;
+use libwebauthn::proto::ctap2::{Ctap2, Ctap2GetInfoResponse};
+use libwebauthn::transport::hid::list_devices;
+use libwebauthn::transport::{Channel as _, Device};
+use text_io::read;
+
+#[path = "../common/mod.rs"]
+mod common;
+
+const TIMEOUT: Duration = Duration::from_secs(10);
+
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+enum Operation {
+    ToggleAlwaysUv,
+    EnableForceChangePin,
+    DisableForceChangePin,
+    SetMinPinLength(Option<u32>),
+    SetMinPinLengthRpids,
+    EnableEnterpriseAttestation,
+}
+
+impl Display for Operation {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Operation::ToggleAlwaysUv => f.write_str("Toggle AlwaysUV"),
+            Operation::EnableForceChangePin => f.write_str("Enable force change pin"),
+            Operation::DisableForceChangePin => f.write_str("Disable force change pin"),
+            Operation::SetMinPinLength(l) => {
+                if let Some(length) = l {
+                    f.write_fmt(format_args!("Set min PIN length. Current length: {length}"))
+                } else {
+                    f.write_str("Set min PIN length.")
+                }
+            }
+            Operation::SetMinPinLengthRpids => f.write_str("Set min PIN length RPIDs"),
+            Operation::EnableEnterpriseAttestation => f.write_str("Enable enterprise attestation"),
+        }
+    }
+}
+
+fn get_supported_options(info: &Ctap2GetInfoResponse) -> Vec<Operation> {
+    let mut configure_ops = vec![];
+    if let Some(options) = &info.options {
+        if options.get("authnrCfg") == Some(&true) && options.get("alwaysUv").is_some() {
+            configure_ops.push(Operation::ToggleAlwaysUv);
+        }
+        if options.get("authnrCfg") == Some(&true) && options.get("setMinPINLength").is_some() {
+            if info.force_pin_change == Some(true) {
+                configure_ops.push(Operation::DisableForceChangePin);
+            } else {
+                configure_ops.push(Operation::EnableForceChangePin);
+            }
+            configure_ops.push(Operation::SetMinPinLength(info.min_pin_length));
+            configure_ops.push(Operation::SetMinPinLengthRpids);
+        }
+        if options.get("ep").is_some() {
+            configure_ops.push(Operation::EnableEnterpriseAttestation);
+        }
+    }
+    configure_ops
+}
+
+#[tokio::main]
+pub async fn main() -> Result<(), Box<dyn Error>> {
+    common::setup_logging();
+
+    let devices = list_devices().await.unwrap();
+    println!("Devices found: {:?}", devices);
+
+    for mut device in devices {
+        println!("Selected HID authenticator: {}", &device);
+        let mut channel = device.channel().await?;
+        channel.wink(TIMEOUT).await?;
+
+        let state_recv = channel.get_ux_update_receiver();
+        tokio::spawn(common::handle_uv_updates(state_recv));
+
+        let info = channel.ctap2_get_info().await?;
+        let options = get_supported_options(&info);
+
+        println!("What do you want to do?");
+        println!();
+        for (idx, op) in options.iter().enumerate() {
+            println!("({idx}) {op}");
+        }
+
+        let idx = common::prompt_index(options.len());
+
+        let mut min_pin_length_rpids = Vec::new();
+        if options[idx] == Operation::SetMinPinLengthRpids {
+            loop {
+                print!("Add RPIDs to list (enter empty string once finished): ");
+                io::stdout().flush().expect("Failed to flush stdout!");
+                let input: String = read!("{}\n");
+                let trimmed = input.trim().to_string();
+                if trimmed.is_empty() {
+                    break;
+                } else {
+                    min_pin_length_rpids.push(trimmed);
+                }
+            }
+        };
+
+        let new_pin_length = if matches!(options[idx], Operation::SetMinPinLength(..)) {
+            loop {
+                print!("New minimum PIN length: ");
+                io::stdout().flush().expect("Failed to flush stdout!");
+                let input: String = read!("{}\n");
+                if let Ok(l) = input.trim().parse::<u64>() {
+                    break l;
+                }
+            }
+        } else {
+            0
+        };
+
+        let result = match options[idx] {
+            Operation::ToggleAlwaysUv => {
+                retry_user_errors!(channel.toggle_always_uv(TIMEOUT)).map(|_| ())
+            }
+            Operation::SetMinPinLengthRpids => retry_user_errors!(
+                channel.set_min_pin_length_rpids(min_pin_length_rpids.clone(), TIMEOUT)
+            )
+            .map(|_| ()),
+            Operation::SetMinPinLength(..) => {
+                retry_user_errors!(channel.set_min_pin_length(new_pin_length, TIMEOUT)).map(|_| ())
+            }
+            Operation::EnableEnterpriseAttestation => {
+                retry_user_errors!(channel.enable_enterprise_attestation(TIMEOUT)).map(|_| ())
+            }
+            Operation::EnableForceChangePin => {
+                retry_user_errors!(channel.force_change_pin(true, TIMEOUT)).map(|_| ())
+            }
+            Operation::DisableForceChangePin => {
+                retry_user_errors!(channel.force_change_pin(false, TIMEOUT)).map(|_| ())
+            }
+        };
+        result.unwrap();
+        println!("Authenticator config done!");
+    }
+
+    Ok(())
+}
